issue #112 rename : getPlanService -> PlanService
