+++ client/app/src/main/java/com/example/witt/data/api/PlanService.kt
@@ -1,15 +1,23 @@
 package com.example.witt.data.api
 
+import com.example.witt.data.model.plan.get_plan.GetPlanListResponse
+import com.example.witt.data.model.plan.get_plan.GetPlanResponse
 import com.example.witt.data.model.plan.join_plan.JoinPlanResponse
 import com.example.witt.data.model.plan.join_plan.OutPlanResponse
 import com.example.witt.data.model.plan.make_plan.request.MakePlanRequest
 import com.example.witt.data.model.plan.make_plan.response.MakePlanResponse
-import retrofit2.http.Body
-import retrofit2.http.DELETE
-import retrofit2.http.PUT
-import retrofit2.http.Path
+import retrofit2.http.*
 
 interface PlanService {
+
+    @GET("/api/trip")
+    suspend fun getPlanList(
+    ): GetPlanListResponse
+
+    @GET("/api/trip/{planId}")
+    suspend fun getPlan(
+        @Path("planId") planId: Int
+    ): GetPlanResponse
 
     @PUT("/api/trip")
     suspend fun makePlan(
