+++ client/app/src/main/java/com/example/witt/data/source/remote/plan/get_plan/GetPlanDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.plan.get_plan
 
-import com.example.witt.data.api.GetPlanService
+import com.example.witt.data.api.PlanService
 import com.example.witt.data.model.plan.get_plan.GetPlanResponse
 import com.example.witt.di.IoDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
@@ -8,7 +8,7 @@ import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class GetPlanDataSourceImpl @Inject constructor(
-    private val getPlanService: GetPlanService,
+    private val getPlanService: PlanService,
     @IoDispatcher private val coroutineDispatcher: CoroutineDispatcher
 ): GetPlanDataSource {
     override suspend fun getPlan(planId: Int): Result<GetPlanResponse>
