+++ client/app/src/main/java/com/example/witt/data/source/remote/plan/get_plan/GetPlanListDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.plan.get_plan
 
-import com.example.witt.data.api.GetPlanService
+import com.example.witt.data.api.PlanService
 import com.example.witt.data.model.plan.get_plan.GetPlanListResponse
 import com.example.witt.di.IoDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
@@ -8,7 +8,7 @@ import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class GetPlanListDataSourceImpl @Inject constructor(
-    private val getPlanService: GetPlanService,
+    private val getPlanService: PlanService,
     @IoDispatcher private val coroutineDispatcher: CoroutineDispatcher
 ): GetPlanListDataSource {
     override suspend fun getPlanList(): Result<GetPlanListResponse>
